Add support after comma for method call
